pyproject.toml
[tool.poetry]
name = "re-factor"
version = "0.1.0"
description = "A github bot which uses LLM for code refactoring"
authors = ["HATIM SAHABBIRBHAI PATRAWALA <ht760280@dal.ca>"]
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "3.11.4"
aiohttp = "^3.8.5"
gidgethub = "^5.3.0"
django = "4.1"
djangorestframework = "^3.14.0"
pre-commit = "^3.4.0"
django-environ = "^0.11.2"
asyncio = "^3.4.3"
mysqlclient = "2.1.1"
django-cors-headers = "^4.2.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
